fix(build): gcc -Werror findings the macOS build cannot see
Linux CI compiles with -Werror (excluded on APPLE), and the first push
tripped it. Reproduced locally with a clang -fsyntax-only -Werror sweep
over all 332 branch TUs (both configurations) and fixed everything a
gcc run reaches:

- expression_executor.cpp: size_t += (cond ? 1 : 0) accumulates an int
  (-Wsign-conversion) -> plain if-increments.
- arithmetic.hpp to_result<bool, floating>: static_cast<bool>(float) is
  an implicit `!= 0` float comparison to gcc (-Wfloat-equal) -> spelled
  with >=/<=, exactly equivalent for every input, NaN included.
- arithmetic.cpp: unused scale_interval removed.
- manager_disk_bootstrap.cpp: scan_batched's [[nodiscard]] result was
  dropped (-Wunused-result); on a failed scan the batch list is now
  cleared and the failure logged -- an empty index rebuild, never a
  silently PARTIAL one.
- test_group_by_node_data.cpp: size_t -> double under -Wconversion.
- python_objects.cpp: int32 -> uint8 in create_decimal (wheel builds
  compile this with -Werror), explicit casts under the existing
  digits10 bound.

ctest 1653/1653 in both configurations.

+++ integration/python/native/python_objects.cpp
@@ -70,7 +70,9 @@ namespace otterbrix {
                         type = logical_type::DOUBLE;
                         return true;
                     }
-                    type = complex_logical_type::create_decimal(width, scale);
+                    // Both fit: width was just bounded by digits10 (18) and scale never exceeds it.
+                    type = complex_logical_type::create_decimal(static_cast<uint8_t>(width),
+                                                                static_cast<uint8_t>(scale));
                     return true;
                 }
                 case py_decimal_exponent_type_t::EXPONENT_INFINITY: {
